voice-server/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
websockets==16.0
aiohttp==3.13.3
redis==7.3.0
twilio==9.10.3
python-dotenv==1.2.2
requests==2.32.5
